data_agent/requirements.txt
langchain
langchain-openai
langchain-community
langchain-tavily
langsmith

python-dotenv
pydantic
matplotlib
seaborn
pandas
pymysql
scikit-learn
IPython
fastapi
uvicorn